requirements.txt
Django==1.6.11
argparse==1.2.1
constants==0.6.0
django-devserver==0.8.0
django-dotenv==1.3.0
django-extensions==1.5.2
django-getenv==1.3.1
django-jsonview==0.4.3
django-nose==1.3
django-tastypie==0.12.1
drest==0.9.12
httplib2==0.9
itsdangerous==0.24
mimeparse==0.1.3
nose==1.3.4
pbr==0.10.8
psycopg2==2.6
python-dateutil==2.4.1
python-mimeparse==0.1.4
requests==2.6.0
six==1.9.0
slumber==0.6.2
sqlparse==0.1.14
stevedore==1.3.0
wsgiref==0.1.2
